# docs/swagger.yaml
basePath: /api/v1
definitions:
  dbmsg.Area:
    properties:
      description:
        $ref: '#/definitions/dbmsg.AreaDescription'
      display_name:
        type: string
      grade:
        type: integer
      id:
        type: integer
    type: object
  dbmsg.AreaDescription:
    properties:
      progress:
        type: string
      target:
        type: string
    type: object
  dbmsg.Circle:
    properties:
      areas:
        items:
          $ref: '#/definitions/dbmsg.Area'
        type: array
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
    type: object
  dbmsg.CreateCircleArea:
    properties:
      description:
        $ref: '#/definitions/dbmsg.AreaDescription'
      display_name:
        type: string
      grade:
        type: integer
    type: object
  handlers.CreateCircleRequest:
    properties:
      areas:
        items:
          $ref: '#/definitions/dbmsg.CreateCircleArea'
        type: array
      created_at:
        type: integer
      description:
        type: string
      user_id:
        type: integer
    type: object
  handlers.CreateCircleResponse:
    properties:
      id:
        type: integer
    type: object
  handlers.GetCirclesRequest:
    type: object
  handlers.GetCirclesResponse:
    properties:
      circles:
        items:
          $ref: '#/definitions/dbmsg.Circle'
        type: array
    type: object
  handlers.GetProfileRequest:
    type: object
  handlers.GetProfileResponse:
    properties:
      display_name:
        type: string
      id:
        type: integer
      login:
        type: string
    type: object
info:
  contact: {}
  description: This is a balance tracker server.
  title: Track Balance API
  version: 0.0.1
paths:
  /api/v1/circle:
    post:
      description: Create circle
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCircleRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateCircleResponse'
      summary: Create circle
      tags:
      - Circles
  /api/v1/circle/{circle_id}:
    delete:
      description: Delete circle
      parameters:
      - description: circle_id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete circle
      tags:
      - Circles
    get:
      parameters:
      - description: circle_id
        in: path
        name: circle_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dbmsg.Circle'
      summary: Get Circle
      tags:
      - Circles
  /api/v1/circles:
    get:
      description: get all circles
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GetCirclesRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetCirclesResponse'
      summary: Get all circles
      tags:
      - Circles
  /api/v1/profile:
    get:
      description: Getting user data
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GetProfileRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetProfileResponse'
      summary: Get Profile
      tags:
      - Profile
swagger: "2.0"